Redactor index.js file
