+++ src/index.js
@@ -1,35 +1,19 @@
 import './stylesheets/style.scss';
-import alooPuri from './images/aloo_puri.jpg';
 
-import createElement from './js/tools';
 import nav from './js/nav';
+import cover from './js/cover'
+import feature from "./js/feature";
 import about from './js/about';
 import footer from './js/footer';
 
 const content = document.getElementById('content');
 
-// Nav
 nav({ appendTo: content });
 
-// cover
-const coverDiv = createElement({ appendTo: content, tag: 'div', cls: 'cover' });
-const coverContent = createElement({ appendTo: coverDiv, tag: 'div', cls: 'cover-content' });
-const coverP1 = createElement({
-  appendTo: coverContent, tag: 'p', cls: 'cover-p1', txt: 'Welcome to ',
-});
-createElement({ appendTo: coverP1, tag: 'span', txt: 'zaika' });
-createElement({
-  appendTo: coverContent, tag: 'p', cls: 'cover-p2', txt: 'Where Food Speaks with your plate!',
-});
+cover({appendTo: content})
 
-// feature-recipe
-const featRecipe = createElement({ appendTo: content, tag: 'div', cls: 'feat-recipe' });
-const featRecipeImg = createElement({ appendTo: featRecipe, cls: 'feat-recipe-img', tag: 'img' });
-featRecipeImg.src = alooPuri;
+feature({appendTo: content})
 
-// About
-const main = createElement({ appendTo: content, tag: 'main' });
-about({ appendTo: main });
+about({appendTo: content})
 
-// Footer
 footer({ appendTo: content });
